requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.0
opencv-python>=4.5.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
omegaconf>=2.3.0
tqdm>=4.60.0
matplotlib>=3.7.0
transformers>=4.30.0
huggingface-hub>=0.20.0
submitit>=1.4.0
